requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
huggingface-hub==0.23.0
scikit-learn==1.4.2
numpy==1.26.4
pillow==10.3.0
timm==1.0.3
torch==2.3.0
torchvision==0.18.0
safetensors==0.4.3
joblib==1.4.0
tensorflow==2.16.1
